ntel® Memory Latency Checker v3.5
To test Memory bandwidth.

##Installation
Copy the mlc binary to any directory on your system
Intel® MLC dynamically links to GNU C library (glibc/lpthread) and this library must be present on the system
Root privileges are required to run this tool as the tool modifies the H/W prefetch control MSR to enable/disable prefetchers for latency and b/w measurements. Refer readme documentation on running without root privileges
MSR driver (not part of the install package) should be loaded. This can typically be done with 'modprobe msr' command if it is not already included.